==> 02_cleaning.ipynb <==
import numpy as np
import pandas as pd
import seaborn as sns
import matplotlib.pyplot as plt
from IPython.display import display # Allows the use of display() for DataFrames

pd.set_option('display.max_columns', None)
plt.style.use(['dark_background'])
%matplotlib inline

reviews_raw = pd.read_csv("reviews_raw.csv")

reviews_raw.head(2)

"""Clean rating"""
def clean_rating(text):
    try:
        rating = int(text[1:][:-4])
    except:
        rating = pd.np.nan
    return rating

print(reviews_raw["rating"].head())
reviews = reviews_raw
reviews["rating"] = reviews_raw["rating"].apply(lambda x: clean_rating(x))
print(reviews["rating"].head())

"""Clean header text"""
def clean_header(header):
    #remove qoute
    text = header.replace('"', '').replace('  ', '')
    if text== "":
        pass
    #remove 'customer review'
    elif text[-15:] == 'customer review':
        text = ""
    elif not text[-1] == '.':
        text += '.'
    return text

reviews.text_header = reviews_raw.text_header.apply(clean_header)

reviews_raw.text.iloc[3][:14]

""" Clean review text"""

def clean_text(text):
    if text== "":
        pass
    #remove 'customer review'
    elif text[0:17] =='✅ Trip Verified |':
        text = text[17:]
    elif text[:20] == '✅ Verified Review | ':
        text = text[20:]
    elif text[:14] == 'Not Verified |':
        text = text[14:]
    elif not text[-1] == '.':
        text += '.'
    return text.replace('"', '').replace('  ', '')

reviews.text = reviews_raw.text.apply(clean_text)

"""Germanwings Flight 9525 accident"""
def days_after_crash(date_published):
    """
    Count days after the accident on 2015.03.25.
    """
    from datetime import datetime
    try:
        days = (datetime.strptime(date_published, "%Y-%m-%d") - datetime(2015, 3, 24)).days
        return days if days>0 else 0
    except:
        return pd.np.nan

reviews["days_after_crash"] = reviews_raw["date_published"].apply(days_after_crash)

print(reviews[["date_published","days_after_crash"]].head())

def add_header(doc):
    text_header = doc[0].replace('"', '').replace('  ', '')
    text = doc[1].replace('"', '').replace('  ', '')
    return text_header + " " + text

reviews["review"] = reviews_raw[["text_header", "text"]].apply(add_header, axis=1)

"""Find germanwings' peers"""
reviews["germanwings"] = reviews["airline_code"] =="germanwings"
germanwings_routes = set(reviews[reviews["germanwings"]]["route"].dropna()) 

peers_set = set(reviews[reviews["route"].isin(germanwings_routes)]["airline_code"])
print(peers_set)

peers = reviews[reviews["airline_code"].isin(peers_set)]

reviews

reviews_text = reviews.groupby("airline_code")["review"].agg({"review": lambda x: "%s" % ' '.join(x)})

#peers.to_csv("peers.csv", index=False)

#reviews_text.to_csv("reviews_text.csv")

#reviews.to_csv("reviews.csv", index=False)
