==> 03_Naive Bayes implementation using scikit-learn.ipynb <==
# Part 1: Load Data

import warnings
warnings.filterwarnings('ignore')
warnings.filterwarnings("ignore",category=DeprecationWarning)
import numpy as np
import pandas as pd
import seaborn as sns
pd.set_option('display.max_columns', 20)


peers = pd.read_csv("peers.csv")
germanwings = peers[peers.airline_code=="germanwings"].drop("germanwings",axis=1)

germanwings.head(1)

## Training a model

from sklearn.model_selection import train_test_split

X_train, X_test, y_train, y_test = train_test_split(germanwings.review, 
                                                    germanwings.recommended, 
                                                    random_state=0)

print('Number of rows in the total set: {}'.format(germanwings.shape[0]))
print('Number of rows in the training set: {}'.format(X_train.shape[0]))
print('Number of rows in the test set: {}'.format(X_test.shape[0]))

from sklearn.feature_extraction.text import CountVectorizer
count_vector = CountVectorizer()
# Fit the training data and then return the matrix
training_data = count_vector.fit_transform(X_train)

# Transform testing data and return the matrix. Note we are not fitting the testing data into the CountVectorizer()
testing_data = count_vector.transform(X_test)

from sklearn.naive_bayes import MultinomialNB
naive_bayes = MultinomialNB()
naive_bayes.fit(training_data, y_train)
predictions = naive_bayes.predict(testing_data)

from sklearn.metrics import accuracy_score, precision_score, recall_score, f1_score
print('Accuracy score: ', format(accuracy_score(predictions,y_test)))
print('Precision score: ', format(precision_score(predictions,y_test)))
print('Recall score: ', format(recall_score(predictions,y_test)))
print('F1 score: ', format(f1_score(predictions,y_test)))

pos_class_prob_sorted = naive_bayes.feature_log_prob_[0, :].argsort()
neg_class_prob_sorted = naive_bayes.feature_log_prob_[1, :].argsort()

print("Top 5 words predicting to recommend:\n" , np.take(count_vector.get_feature_names(), pos_class_prob_sorted[:5]))
print("\nTop emotions = emotion_analyzer_documents(pressed_reviews)
emotions5 words predicting to not recommend:\n", np.take(count_vector.get_feature_names(), neg_class_prob_sorted[:5]))
